requirements.txt
httpx==0.27.0
openai==1.37.0
Pillow==11.2.1
pdfix-sdk==8.4.3
requests==2.32.4
types-requests==2.32.0.20250515
